# api/openapi.yaml
openapi: 3.0.3
info:
  title: podman-api
  version: 1.0.0
  description: |
    REST wrapper around Podman's libpod API. Manages parameterised pod
    instances across a fleet of remote Podman hosts via SSH-tunneled
    sockets. See README.md for the deployment story.
servers:
  - url: http://127.0.0.1:8080
    description: local loopback (default)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Argon2id-hashed bearer tokens defined in `auth/keys.yaml`. The
        token's `scopes` must include the scope listed on each operation.
  parameters:
    HostID:
      name: host
      in: path
      required: true
      schema: {type: string, pattern: '^[a-z0-9][a-z0-9-]{0,38}[a-z0-9]$'}
...
  schemas:
...
    PortMapping:
      type: object
      required: [host_port, container_port, protocol]
      properties:
        host_ip: {type: string}
        host_port: {type: integer}
        container_port: {type: integer}
        protocol: {type: string, enum: [tcp, udp]}
        pod: {type: string}
        container: {type: string}
...
    JobAccepted:
      type: object
      required: [job_id]
      properties:
        job_id: {type: string}
...
  responses:
    Unauthorized:
      description: missing or invalid bearer token
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Forbidden:
      description: token lacks required scope
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...
    NotFound:
      description: unknown host, template, or instance
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...
security:
  - bearerAuth: []
paths:
...
  /hosts:
    get:
      tags: [hosts]
      summary: list configured hosts with reachability and instance count
      responses:
        '200':
          description: list
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/Host'}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
      x-required-scope: hosts:read
  /hosts/{host}:
    get:
      tags: [hosts]
      summary: one host's reachability and instance count
      parameters: [{$ref: '#/components/parameters/HostID'}]
      responses:
        '200':
          description: host
          content: {application/json: {schema: {$ref: '#/components/schemas/Host'}}}
        '404': {$ref: '#/components/responses/NotFound'}
      x-required-scope: hosts:read
...
  /hosts/{host}/ports-in-use:
    get:
      tags: [hosts]
      summary: every host port currently bound by a managed pod
      parameters: [{$ref: '#/components/parameters/HostID'}]
      responses:
        '200':
          description: ports
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/PortMapping'}
      x-required-scope: hosts:read
  /hosts/{host}/rename:
    post:
      tags: [hosts]
      summary: rename a host's id in the config store
      description: >-
        Changes a host's id in the config file and the control-plane state
        store, and makes the new id live immediately — no restart or SIGHUP.
        Fails if the new id is already in use, if the host has backups (see
        409), or when the HostFileRenamer is not wired (501 not_implemented).
      parameters: [{$ref: '#/components/parameters/HostID'}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [new_id]
              properties:
                new_id:
                  type: string
                  pattern: '^[a-z0-9][a-z0-9-]{0,38}[a-z0-9]$'
                  description: the new id for the host
      responses:
        '200':
          description: host renamed
          content:
            application/json:
              schema:
                type: object
                properties:
                  old_id: {type: string}
                  new_id: {type: string}
        '400':
          description: >-
            invalid request body: new_id is required, new_id must differ from
            the current host id, or new_id does not match the allowed pattern
          content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
        '404':
          description: unknown host
          content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
        '409':
          description: >-
            two distinct refusals share this status; tell them apart by the
            error code. host_already_exists — new_id is already in use by
            another host, or the state store already holds specs, host secrets
            or backups under it. host_has_backups — the host being renamed has
            backup rows; their blob keys are NOT re-keyed by a rename, so
            renaming would leave those backups unreachable through the restore
            route. Delete the backups first, or wait for the blob re-key
            follow-up.
          content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
        '501':
          description: host rename is not configured on this server (HostFileRenamer not wired)
          content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
      x-required-scope: hosts:write
  /templates:
    get:
      tags: [templates]
      summary: list templates from the store
      responses:
        '200':
          description: list
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/Template'}}
      x-required-scope: templates:read
    post:
      tags: [templates]
      summary: create a new template in the store
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/TemplateInput'}}}
      responses:
        '201':
          description: created
          content: {application/json: {schema: {$ref: '#/components/schemas/Template'}}}
        '400':
          description: invalid body or invalid template (bad id, meta, or render)
          content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
        '409':
          description: a template with this id already exists (template_already_exists)
          content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
      x-required-scope: templates:write
  /templates/{id}:
    get:
      tags: [templates]
      summary: one template's metadata
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string, pattern: '^[a-z0-9][a-z0-9-]{0,38}[a-z0-9]$'}
      responses:
        '200':
          description: template
          content: {application/json: {schema: {$ref: '#/components/schemas/Template'}}}
        '404': {$ref: '#/components/responses/NotFound'}
      x-required-scope: templates:read
    put:
      tags: [templates]
      summary: update an existing template (id is taken from the path)
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string, pattern: '^[a-z0-9][a-z0-9-]{0,38}[a-z0-9]$'}
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/TemplateInput'}}}
      responses:
        '200':
          description: updated
          content: {application/json: {schema: {$ref: '#/components/schemas/Template'}}}
        '400':
          description: invalid body or invalid template (bad meta or render)
          content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
        '404': {$ref: '#/components/responses/NotFound'}
      x-required-scope: templates:write
    delete:
      tags: [templates]
      summary: delete a template
      description: >-
        Fails with 409 (template_in_use) when instances are still deployed from
        this template, unless force=true is passed.
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string, pattern: '^[a-z0-9][a-z0-9-]{0,38}[a-z0-9]$'}
        - name: force
          in: query
          required: false
          description: delete even when instances are still deployed from this template
          schema: {type: boolean, default: false}
      responses:
        '204': {description: removed}
        '404': {$ref: '#/components/responses/NotFound'}
        '409':
          description: instances are still deployed from this template (template_in_use)
          content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
      x-required-scope: templates:write
  /templates/{id}/clone:
    post:
      tags: [templates]
      summary: clone an existing template under a new id
      parameters:
        - name: id
          in: path
          required: true
          description: id of the source template to clone
          schema: {type: string, pattern: '^[a-z0-9][a-z0-9-]{0,38}[a-z0-9]$'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [new_id]
              properties:
                new_id: {type: string, pattern: '^[a-z0-9][a-z0-9-]{0,38}[a-z0-9]$'}
      responses:
        '201':
          description: created
          content: {application/json: {schema: {$ref: '#/components/schemas/Template'}}}
        '400':
          description: invalid body or invalid new_id
          content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
        '404': {$ref: '#/components/responses/NotFound'}
        '409':
          description: a template with new_id already exists (template_already_exists)
          content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
      x-required-scope: templates:write
  /templates/{id}/render:
    get:
      tags: [templates]
      summary: render a template with query-string parameters (preview only)
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string, pattern: '^[a-z0-9][a-z0-9-]{0,38}[a-z0-9]$'}
      responses:
        '200':
          description: rendered YAML
          content: {text/yaml: {schema: {type: string}}}
      x-required-scope: templates:read
  /registry/repos:
    get:
      tags: [registry]
      summary: list every repository in the configured registry (Registry v2 catalog)
      description: >-
        404 when no registry client is configured (-registry-* flags), the
        same "absent, not disabled" shape as every other route under
        /registry.
      responses:
        '200':
          description: repository names
          content:
            application/json:
              schema: {type: array, items: {type: string}}
        '404':
          description: registry browsing is not configured
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
        '502': {$ref: '#/components/responses/UpstreamError'}
      x-required-scope: instances:read
  /registry/repos/{repo}:
    get:
      tags: [registry]
      summary: list a repository's tags, or resolve a single tag/digest's manifest
      description: >-
        With no query string, lists every tag in `repo` grouped by content
        digest. With `?manifest=<ref>`, resolves that one tag or digest
        instead and returns its manifest. Both are served by this one route
        because Go 1.22's `{repo...}` multi-segment wildcard — required so a
        "/"-containing repo name (e.g. `iotready/engine`) is addressable at
        all — must be the mux's last path segment, ruling out a separate
        `.../manifests/{ref}` route. `repo` itself may contain further `/`
        segments (e.g. `iotready/engine`). 404 when no registry client is
        configured, same as GET /registry/repos. A genuine registry 404 (no
        such repo/tag/digest) and a registry that could not be reached are
        never collapsed into the same status: the former is 404 not_found,
        the latter is 502 registry_unreachable.
      parameters:
        - name: repo
          in: path
          required: true
          description: repository name; may contain additional "/" segments
          schema: {type: string}
        - name: manifest
          in: query
          required: false
          description: a tag or a digest (algorithm:hex); when set, returns that ref's manifest instead of the tag listing
          schema: {type: string}
      responses:
        '200':
          description: tag groups (no `manifest` query param), or one manifest (`manifest` set)
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items: {$ref: '#/components/schemas/RegistryTagGroup'}
                  - $ref: '#/components/schemas/RegistryManifest'
        '400':
          description: invalid repo name, or an invalid `manifest` value (must be a tag or a digest)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
        '404':
          description: registry browsing is not configured, or (once configured) no such repo/tag/digest (code not_found)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
        '502': {$ref: '#/components/responses/UpstreamError'}
      x-required-scope: instances:read
  /registry/prune:
    post:
      tags: [registry]
      summary: enqueue an on-demand registry prune run (#219)
      description: >-
        Classifies every manifest in the configured registry against live
        fleet inventory and stored specs, then unlinks the ones that are
        unreferenced and safely recreatable, failing closed on any doubt
        (e.g. a truncated inventory sweep). Whether the run is dry-run, its
        deletion policy, and whether blob GC (Stage B) runs are all fixed
        server-side flags (-registry-prune-*) — the request body is empty,
        so a caller cannot turn a dry-run deployment into a deleting one over
        HTTP. Read the run's progress via GET /jobs/{job_id}. 404 when the
        feature is not enabled (-registry-prune-enabled), the same
        "absent, not disabled" shape as the browse routes above. 409 when a
        run is already in flight — the scheduler's own guard against two
        runs classifying (and mutating) the same catalog concurrently; this
        is deliberately not a 202 carrying the in-flight job's id, since
        "your run started" and "someone else's run is still going" are
        different answers.
      responses:
        '202':
          description: prune job accepted
          content:
            application/json:
              schema: {$ref: '#/components/schemas/JobAccepted'}
        '404':
          description: registry prune is not configured (-registry-prune-enabled)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
        '409':
          description: a prune run is already in flight (registryprune.ErrRunInFlight)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
      x-required-scope: instances:write
  /hosts/{host}/secrets:
    get:
      tags: [secrets]
      summary: list per-host secrets known to podman
      parameters: [{$ref: '#/components/parameters/HostID'}]
      responses:
        '200':
          description: list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    name: {type: string}
                    created_at: {type: string, format: date-time}
      x-required-scope: secrets:read
  /hosts/{host}/secrets/{name}:
    put:
      tags: [secrets]
      summary: create or rotate a per-host secret
      parameters:
        - {$ref: '#/components/parameters/HostID'}
        - {$ref: '#/components/parameters/SecretName'}
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/SecretValue'}}}
      responses:
        '204': {description: stored}
        '400': {$ref: '#/components/responses/BadRequest'}
      x-required-scope: secrets:write
    delete:
      tags: [secrets]
      summary: remove a per-host secret (idempotent)
      parameters:
        - {$ref: '#/components/parameters/HostID'}
        - {$ref: '#/components/parameters/SecretName'}
      responses:
        '204': {description: removed}
      x-required-scope: secrets:write
  /hosts/{host}/instances:
    get:
      tags: [instances]
      summary: list managed instances on a host (optionally filtered by template)
      parameters:
        - {$ref: '#/components/parameters/HostID'}
        - {$ref: '#/components/parameters/TemplateOptional'}
      responses:
        '200':
          description: list
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/Observed'}}
      x-required-scope: instances:read
    post:
      tags: [instances]
      summary: create a new instance (fails if it already exists)
      parameters:
        - {$ref: '#/components/parameters/HostID'}
        - {$ref: '#/components/parameters/SkipPull'}
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/ApplyRequest'}}}
      responses:
        '201':
          description: created
          content: {application/json: {schema: {$ref: '#/components/schemas/Observed'}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '423': {$ref: '#/components/responses/HostDraining'}
        '502': {$ref: '#/components/responses/UpstreamError'}
      x-required-scope: instances:write
  /hosts/{host}/instances/{template}/{slug}:
    get:
      tags: [instances]
      summary: observed shape of one instance
      parameters:
        - {$ref: '#/components/parameters/HostID'}
        - {$ref: '#/components/parameters/TemplateID'}
        - {$ref: '#/components/parameters/Slug'}
      responses:
        '200':
          description: observed
          content: {application/json: {schema: {$ref: '#/components/schemas/Observed'}}}
        '404': {$ref: '#/components/responses/NotFound'}
      x-required-scope: instances:read
    put:
      tags: [instances]
      summary: create or replace an instance (idempotent)
      parameters:
        - {$ref: '#/components/parameters/HostID'}
        - {$ref: '#/components/parameters/TemplateID'}
        - {$ref: '#/components/parameters/Slug'}
        - {$ref: '#/components/parameters/SkipPull'}
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/ApplyRequest'}}}
      responses:
        '200':
          description: applied
          content: {application/json: {schema: {$ref: '#/components/schemas/Observed'}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '423': {$ref: '#/components/responses/HostDraining'}
        '502': {$ref: '#/components/responses/UpstreamError'}
      x-required-scope: instances:write
    delete:
      tags: [instances]
      summary: remove an instance and optionally its volumes/secrets
      parameters:
        - {$ref: '#/components/parameters/HostID'}
        - {$ref: '#/components/parameters/TemplateID'}
        - {$ref: '#/components/parameters/Slug'}
        - {name: prune_volumes, in: query, schema: {type: boolean}}
        - {name: prune_secrets, in: query, schema: {type: boolean}}
      responses:
        '204': {description: removed}
        '404': {$ref: '#/components/responses/NotFound'}
      x-required-scope: instances:write
  /hosts/{host}/instances/{template}/{slug}/start:
    post:
      tags: [lifecycle]
      summary: start a stopped instance
      parameters:
        - {$ref: '#/components/parameters/HostID'}
        - {$ref: '#/components/parameters/TemplateID'}
        - {$ref: '#/components/parameters/Slug'}
      responses:
        '204': {description: started}
        '404': {$ref: '#/components/responses/NotFound'}
      x-required-scope: instances:write
  /hosts/{host}/instances/{template}/{slug}/stop:
    post:
      tags: [lifecycle]
      summary: stop a running instance
      parameters:
        - {$ref: '#/components/parameters/HostID'}
        - {$ref: '#/components/parameters/TemplateID'}
        - {$ref: '#/components/parameters/Slug'}
      responses:
        '204': {description: stopped}
      x-required-scope: instances:write
  /hosts/{host}/instances/{template}/{slug}/restart:
    post:
      tags: [lifecycle]
      summary: restart an instance
      parameters:
        - {$ref: '#/components/parameters/HostID'}
        - {$ref: '#/components/parameters/TemplateID'}
        - {$ref: '#/components/parameters/Slug'}
      responses:
        '204': {description: restarted}
      x-required-scope: instances:write
  /hosts/{host}/instances/{template}/{slug}/upgrade:
    post:
      tags: [lifecycle]
      summary: replace an instance's image
      parameters:
        - {$ref: '#/components/parameters/HostID'}
        - {$ref: '#/components/parameters/TemplateID'}
        - {$ref: '#/components/parameters/Slug'}
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/UpgradeRequest'}}}
      responses:
        '200':
          description: upgraded
          content: {application/json: {schema: {$ref: '#/components/schemas/Observed'}}}
        '502': {$ref: '#/components/responses/UpstreamError'}
      x-required-scope: instances:write
  /hosts/{host}/instances/{template}/{slug}/upgrade-image:
    post:
      tags: [lifecycle]
      summary: image-only upgrade reusing the instance's stored spec and sealed secrets
      description: >-
        Reuses the instance's stored parameters and sealed per-instance secrets,
        overriding only the image, then re-applies with replace. Unlike
        /upgrade this requires no secrets in the request — the sealed secrets are
        unrecoverable plaintext and never need to be re-supplied — so an
        unattended bearer-auth loop can roll many instances to a new image
        without an operator login. A per-instance secret the template added after
        deploy does not block the upgrade (the missing secret was already
        missing).
      parameters:
        - {$ref: '#/components/parameters/HostID'}
        - {$ref: '#/components/parameters/TemplateID'}
        - {$ref: '#/components/parameters/Slug'}
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/UpgradeImageRequest'}}}
      responses:
        '200':
          description: upgraded
          content: {application/json: {schema: {$ref: '#/components/schemas/Observed'}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '502': {$ref: '#/components/responses/UpstreamError'}
      x-required-scope: instances:write
...
  /hosts/{host}/bulk:
    post:
      tags: [bulk]
      summary: run lifecycle ops against many instances on one host
      parameters: [{$ref: '#/components/parameters/HostID'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/BulkRequest'}}}
      responses:
        '200':
          description: per-op results (response is 200 even when individual ops fail)
          content: {application/json: {schema: {$ref: '#/components/schemas/BulkResult'}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '413':
          description: more than 100 ops in a single request
          content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
      x-required-scope: instances:write
...
  /hosts/{host}/instances/{template}/{slug}/volumes:
    get:
      tags: [volumes]
      summary: volumes belonging to one instance
      parameters:
        - {$ref: '#/components/parameters/HostID'}
        - {$ref: '#/components/parameters/TemplateID'}
        - {$ref: '#/components/parameters/Slug'}
      responses:
        '200':
          description: volumes
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/Volume'}}
      x-required-scope: instances:read
...