# internal/api/router.go
package api

import (
	"net/http"

	apispec "github.com/iotready/podman-api/api"
	"github.com/iotready/podman-api/internal/auth"
	"github.com/iotready/podman-api/internal/imgregistry"
	"github.com/iotready/podman-api/internal/instance"
	"github.com/iotready/podman-api/internal/store"
)

// HostFileRenamer rewrites a host's id in its hosts/*.yaml config file,
// returning the file's path and its original content (for revert-on-failure).
// Implemented by config.HostsDir.
type HostFileRenamer interface {
	RenameHostFile(oldID, newID string) (path string, original []byte, err error)
}

// NewRouter builds the full HTTP handler tree.
// audit is an optional middleware applied around auth-guarded handlers; pass nil for no-op.
// metricsHandler is an optional handler mounted at GET /metrics; pass nil to omit the endpoint.
// canceller is an optional JobCanceller for the POST /jobs/{id}/cancel route; pass nil when the job runner is not wired.
// version is the server's release string (e.g. "v1.0.16", "dev"); included in the /mcp discovery document.
// registryClient is an optional imgregistry.Client for the GET /registry/... routes; pass nil to disable registry browsing (routes return 404).
// opts carry the optional dependencies added after this signature settled; see RouterOption.
func NewRouter(svc *instance.Service, jobs store.JobStore, keys *auth.KeyStore, audit func(http.Handler) http.Handler, metricsHandler http.Handler, canceller JobCanceller, version string, registryClient imgregistry.Client, opts ...RouterOption) http.Handler {
	mux := http.NewServeMux()
	h := &handlers{svc: svc, jobs: jobs, canceller: canceller, registry: registryClient}
	for _, o := range opts {
		o(h)
	}

	if audit == nil {
		audit = func(h http.Handler) http.Handler { return h }
	}

	// Process endpoints (no auth).
	mux.HandleFunc("GET /healthz", func(w http.ResponseWriter, _ *http.Request) {
		w.WriteHeader(http.StatusOK)
		_, _ = w.Write([]byte(`{"status":"ok"}`))
	})
	mux.HandleFunc("GET /openapi.yaml", func(w http.ResponseWriter, _ *http.Request) {
		w.Header().Set("Content-Type", "text/yaml; charset=utf-8")
		w.WriteHeader(http.StatusOK)
		_, _ = w.Write(apispec.Spec)
	})
	mux.HandleFunc("GET /mcp", mcpDiscoveryHandler(version))
	mux.HandleFunc("GET /agent-docs", agentDocsHandler)

	if metricsHandler != nil {
		mux.Handle("GET /metrics", metricsHandler)
	}

	guard := func(scope string, h http.Handler) http.Handler {
		return auth.New(keys, scope)(audit(h))
	}

	// Hosts (read).
	mux.Handle("GET /hosts", guard("hosts:read", http.HandlerFunc(h.listHosts)))
	mux.Handle("GET /hosts/{host}", guard("hosts:read", http.HandlerFunc(h.getHost)))
	mux.Handle("GET /hosts/{host}/healthz", guard("hosts:read", http.HandlerFunc(h.hostHealthz)))
	mux.Handle("GET /hosts/{host}/ports-in-use", guard("hosts:read", http.HandlerFunc(h.portsInUse)))

	// Host rename. 501 when no HostFileRenamer is wired (WithHostRenamer).
	mux.Handle("POST /hosts/{host}/rename", guard("hosts:write", http.HandlerFunc(h.renameHost)))

	// Templates (catalog CRUD + clone).
	mux.Handle("GET /templates", guard("templates:read", http.HandlerFunc(h.listTemplates)))
	mux.Handle("POST /templates", guard("templates:write", http.HandlerFunc(h.createTemplate)))
	mux.Handle("GET /templates/{id}", guard("templates:read", http.HandlerFunc(h.getTemplate)))
	mux.Handle("PUT /templates/{id}", guard("templates:write", http.HandlerFunc(h.updateTemplate)))
	mux.Handle("DELETE /templates/{id}", guard("templates:write", http.HandlerFunc(h.deleteTemplate)))
	mux.Handle("POST /templates/{id}/clone", guard("templates:write", http.HandlerFunc(h.cloneTemplate)))
	mux.Handle("GET /templates/{id}/render", guard("templates:read", http.HandlerFunc(h.renderTemplate)))

	// Registry (read-only browsing of a Docker Registry v2 endpoint). 404 when no
	// registry client is configured.
	//
	// Registry v2 repo names legally contain "/" (e.g. "iotready/engine"), so
	// {repo} (one path segment) can't address every real repo — {repo...}
	// (Go 1.22's multi-segment wildcard) is required. That wildcard must be
	// the mux's last segment, so a separate .../manifests/{ref} route can no
	// longer follow it; the manifest lookup is folded into the same route via
	// a "?manifest=<ref>" query parameter instead (see getRepoOrManifest).
	mux.Handle("GET /registry/repos", guard("instances:read", http.HandlerFunc(h.listRepos)))
	mux.Handle("GET /registry/repos/{repo...}", guard("instances:read", http.HandlerFunc(h.getRepoOrManifest)))

	// On-demand registry prune. Write-scoped because it enqueues a job that
	// deletes manifests (unless the server is configured dry-run), and 404 when
	// the feature is not wired — the same "absent, not disabled" shape as the
	// browse routes above.
	mux.Handle("POST /registry/prune", guard("instances:write", http.HandlerFunc(h.enqueueRegistryPrune)))

	// Host secrets.
	mux.Handle("GET /hosts/{host}/secrets", guard("secrets:read", http.HandlerFunc(h.listSecrets)))
	mux.Handle("PUT /hosts/{host}/secrets/{name}", guard("secrets:write", http.HandlerFunc(h.putSecret)))
	mux.Handle("DELETE /hosts/{host}/secrets/{name}", guard("secrets:write", http.HandlerFunc(h.deleteSecret)))

	// Instances.
	mux.Handle("GET /hosts/{host}/instances", guard("instances:read", http.HandlerFunc(h.listInstances)))
	mux.Handle("GET /hosts/{host}/instances/{template}/{slug}", guard("instances:read", http.HandlerFunc(h.getInstance)))
	mux.Handle("POST /hosts/{host}/instances", guard("instances:write", http.HandlerFunc(h.createInstance)))
	mux.Handle("PUT /hosts/{host}/instances/{template}/{slug}", guard("instances:write", http.HandlerFunc(h.applyInstance)))
	mux.Handle("DELETE /hosts/{host}/instances/{template}/{slug}", guard("instances:write", http.HandlerFunc(h.deleteInstance)))

	// Lifecycle.
	mux.Handle("POST /hosts/{host}/instances/{template}/{slug}/start", guard("instances:write", http.HandlerFunc(h.startInstance)))
	mux.Handle("POST /hosts/{host}/instances/{template}/{slug}/stop", guard("instances:write", http.HandlerFunc(h.stopInstance)))
	mux.Handle("POST /hosts/{host}/instances/{template}/{slug}/restart", guard("instances:write", http.HandlerFunc(h.restartInstance)))
	mux.Handle("POST /hosts/{host}/instances/{template}/{slug}/upgrade", guard("instances:write", http.HandlerFunc(h.upgradeInstance)))
	mux.Handle("POST /hosts/{host}/instances/{template}/{slug}/upgrade-image", guard("instances:write", http.HandlerFunc(h.upgradeImageInstance)))
	mux.Handle("PATCH /hosts/{host}/instances/{template}/{slug}/parameters", guard("instances:write", http.HandlerFunc(h.patchInstanceParameters)))
	mux.Handle("PATCH /hosts/{host}/instances/{template}/{slug}/secrets", guard("instances:write", http.HandlerFunc(h.patchInstanceSecrets)))
	mux.Handle("POST /hosts/{host}/instances/{template}/{slug}/rename", guard("instances:write", http.HandlerFunc(h.renameInstance)))

	// Bulk lifecycle operations against many instances on one host.
	mux.Handle("POST /hosts/{host}/bulk", guard("instances:write", http.HandlerFunc(h.bulk)))

	// Logs.
	mux.Handle("GET /hosts/{host}/instances/{template}/{slug}/logs", guard("instances:read", http.HandlerFunc(h.logsInstance)))

	// Volumes.
	mux.Handle("GET /hosts/{host}/instances/{template}/{slug}/volumes", guard("instances:read", http.HandlerFunc(h.instanceVolumes)))
	mux.Handle("DELETE /hosts/{host}/volumes/{name}", guard("instances:write", http.HandlerFunc(h.deleteVolume)))

	// Jobs (read). 501 when the store is disabled.
	mux.Handle("GET /jobs", guard("jobs:read", http.HandlerFunc(h.listJobs)))
	mux.Handle("GET /jobs/{id}", guard("jobs:read", http.HandlerFunc(h.getJob)))
	mux.Handle("POST /jobs/{id}/cancel", guard("instances:write", http.HandlerFunc(h.cancelJob)))

	// Migrate enqueues a job; 501 when the store is disabled.
	mux.Handle("POST /migrate", guard("instances:write", http.HandlerFunc(h.migrate)))

	// Backups: enqueue/list per instance; restore/delete per backup (#66).
	mux.Handle("POST /hosts/{host}/instances/{template}/{slug}/backup", guard("instances:write", http.HandlerFunc(h.postBackup)))
	mux.Handle("GET /hosts/{host}/instances/{template}/{slug}/backups", guard("instances:read", http.HandlerFunc(h.listBackups)))
	mux.Handle("POST /backups/{id}/restore", guard("instances:write", http.HandlerFunc(h.postRestore)))
	mux.Handle("DELETE /backups/{id}", guard("instances:write", http.HandlerFunc(h.deleteBackup)))

	// Point-in-time restore: roll a live instance back to a chosen timestamp via
	// the injected restore initContainer (commercial Litestream). Distinct from
	// the per-backup tarball restore above (#25 seam).
	mux.Handle("POST /hosts/{host}/instances/{template}/{slug}/restore", guard("instances:write", http.HandlerFunc(h.postPITRRestore)))

	// Evacuate enqueues a parent job that fans out child migrate jobs; 501 when the store is disabled.
	mux.Handle("POST /evacuate", guard("instances:write", http.HandlerFunc(h.evacuate)))

	// Evacuate plan is a read-only dry-run: resolve the map and run live
	// destination preflight checks, returning a per-move report. No job, no
	// mutation; 501 when the store is disabled.
	mux.Handle("POST /evacuate/plan", guard("instances:read", http.HandlerFunc(h.evacuatePlan)))

	return mux
}

// handlers holds per-request dependencies. Each method is a thin adapter
// around svc.
type handlers struct {
	svc           *instance.Service
	jobs          store.JobStore
	canceller     JobCanceller
	registry      imgregistry.Client
	pruner        RegistryPruner
	hostRenamer   HostFileRenamer
	hostsReloader func() error
}

// RouterOption supplies an optional dependency to NewRouter. Options exist
// because NewRouter's positional signature already has eight parameters and is
// called from ~26 places; a new optional dependency should not touch every one
// of them.
type RouterOption func(*handlers)

// WithRegistryPruner enables POST /registry/prune. Without it the route 404s,
// exactly as the browse routes do without a registry client.
func WithRegistryPruner(p RegistryPruner) RouterOption {
	return func(h *handlers) { h.pruner = p }
}

// WithHostRenamer enables POST /hosts/{host}/rename. Without it the route
// responds 501, the same "absent, not disabled" shape as the registry browse
// routes without a client.
func WithHostRenamer(r HostFileRenamer) RouterOption {
	return func(h *handlers) { h.hostRenamer = r }
}

// WithHostsReloader supplies "re-read hosts/*.yaml and make it live
// everywhere", i.e. exactly what SIGHUP does. The rename handler calls it after
// a successful rename so the podman client's host map and the server's
// background-poller host list follow the new id, instead of working the old one
// until the next SIGHUP.
//
// Unlike WithHostRenamer its absence is not a 501: the reload runs AFTER the
// rename has been committed to both the config file and the store, so a missing
// or failing reloader is logged, never a failed request.
func WithHostsReloader(reload func() error) RouterOption {
	return func(h *handlers) { h.hostsReloader = reload }
}